fix: need to restore logstream on follower
Leader to follower transition had issues, because Leader closed
logstream (in context logstream was null) Logstream NEED to be reseted
for leaders, to overcome it we just reset it also on followers. Ideally
we rework the transition mechanism. So we can figure the transitions as
they are which contains the steps we really need to do.